+++ python/gravel/__init__.py
@@ -178,7 +178,7 @@ try:
 except ImportError:
     pass  # OSM support not compiled in
 
-__version__ = "2.2.0"
+__version__ = "2.2.1"
 
 __all__ = [
     # Core